Minor update to make ESTIA consistant

+++ Model/essBuild/essVariables.cxx
@@ -404,8 +404,6 @@ EssBeamLinesVariables(FuncDataBase& Control)
   return;
 }
 
-
-
 void
 EssFlightLineVariables(FuncDataBase& Control)
   /*!
@@ -431,9 +429,9 @@ EssFlightLineVariables(FuncDataBase& Control)
   // Full height = TopFlyTotalHeight
   Control.addVariable("TopAFlightHeight", 2.9);     
   Control.addVariable("TopAFlightWidth", 10.7);     // Full width
-  Control.addVariable("TopAFlightNLiner", 1);      // Liner
-  Control.addVariable("TopAFlightLinerThick1", 0.3);      // Liner
-  Control.addVariable("TopAFlightLinerMat1","Aluminium");      // Liner
+  Control.addVariable("TopAFlightNLiner",1);      
+  Control.addVariable("TopAFlightLinerThick1",0.3);      
+  Control.addVariable("TopAFlightLinerMat1","Aluminium");      
 
   // LOWER flight lines
   // B FLIGHT CORRECTED
@@ -469,13 +467,13 @@ EssFlightLineVariables(FuncDataBase& Control)
   std::vector<double> TopAFlightWedgeTheta;
   TopAFlightWedgeTheta.push_back(t1);
   for (size_t i=1; i<=3; i++)
-    TopAFlightWedgeTheta.push_back(t1-dt1*i);
+    TopAFlightWedgeTheta.push_back(t1-dt1*static_cast<double>(i));
 
   // central wedge: Rickard Holmberg slide 14
   TopAFlightWedgeTheta.push_back(-2.8);
   TopAFlightWedgeTheta.push_back(t2);
   for (size_t i=1; i<=8; i++)
-    TopAFlightWedgeTheta.push_back(t2-dt2*i);
+    TopAFlightWedgeTheta.push_back(t2-dt2*static_cast<double>(i));
 
   double xstep(0);
   double ystep(0);
@@ -520,7 +518,7 @@ EssFlightLineVariables(FuncDataBase& Control)
   const double t3 = (9.1+dt2*8.0)-180.0; 
   TopBFlightWedgeTheta.push_back(t3);
   for (size_t i=1;i<=8;i++)
-    TopBFlightWedgeTheta.push_back(t3-dt2*i);
+    TopBFlightWedgeTheta.push_back(t3-dt2*static_cast<double>(i));
 
   // email from Rickard Holmberg 15 Sep, slide 6
   const double t4 = -15.45-180;
